Implement Kernel. Improve Kernel's func.

+++ MenthaAssembly/Media/Imaging/Filters/ImageFilterArgs.cs
@@ -4,13 +4,13 @@
     {
         public bool Handled { set; get; }
 
-        public int TokenA { set; get; }
+        public float TokenA { set; get; }
 
-        public int TokenR { set; get; }
+        public float TokenR { set; get; }
 
-        public int TokenG { set; get; }
+        public float TokenG { set; get; }
 
-        public int TokenB { set; get; }
+        public float TokenB { set; get; }
 
         // More kernel need?
         // Need List :
